Creating base for widget system

+++ Sparkle/includes/Structure/spk_json_parser.h
@@ -263,7 +263,7 @@ namespace spk
 		{
 			return (std::get<std::string>(get(key)));
 		}
-		float getFloatingPoint(std::string key)
+		double getFloatingPoint(std::string key)
 		{
 			return (std::get<double>(get(key)));
 		}
